Refactor migration package structure and improve database migration process
